wrap mv in a file test

+++ ci/do_circle_ci.sh
@@ -2,9 +2,11 @@
 
 set -e
 
-# bazel uses jgit internally and the default circle-ci .gitconfig says to
-# convert https://github.com to ssh://git@github.com, which jgit does not support.
-mv ~/.gitconfig ~/.gitconfig_save
+if [[ -f ~/.gitconfig ]]; then
+  # bazel uses jgit internally and the default circle-ci .gitconfig says to
+  # convert https://github.com to ssh://git@github.com, which jgit does not support.
+  mv ~/.gitconfig ~/.gitconfig_save
+fi
 
 export ENVOY_SRCDIR="$(pwd)"
